Convert Tokyo Mirage Resolution pack (and more) to V5
Still don't know what to do with Anti-Aliasing packs but those will come at a later date.

I did a bunch of stuff with the resolution pack with ultrawide resolution specific options, but none of the packs are actually tested. Should work, but testing is done in the last phase when I'll be open for community feedback and testing.